upload item with image
