+++ app/router.js
@@ -13,6 +13,7 @@ Router.map(function() {
       this.route('index', { path: '/'});
       this.resource('items', function(){
         this.route('new');
+        this.route('add_item');
         this.resource('item', { path: '/:item_id'});
       });
     });
