+++ app/templates/item.hbs
@@ -6,8 +6,7 @@
 </p>
 
 <b>IMAGES</b>
-
+<br/>
 {{#each images}}
-  <br/>
   <img {{bind-attr src="thumbImageUrl"}}>
 {{/each}}
