+++ app/templates/items.hbs
@@ -1,1 +1,2 @@
+<input id="images_identifiers" name="imageIdentifiers" type="hidden" value="">
 {{outlet}}
