+++ app/templates/items/new.hbs
@@ -10,30 +10,11 @@
   </div>
 
   <div>
-    <ul class='file_names'>
+    <ul class='file_preview'>
     </ul>
   </div>
 
   <div class='sceneUpBtn'></div>
 
-  <input id="images_identifiers" name="image[identifiers]" type="hidden" value="">
-
-  <p>
-    <label>Donor description</label>
-    {{input type="text" value=donorDescription name="donorDescription" }}
-    {{input type="hidden" value=offerId name="offerId" }}
-  </p>
-
-  <p>
-    Condition? <br/>
-
-    {{#each condition in itemConditions}}
-      <label>
-        {{view 'radio-button' name="selectionTest" selectionBinding="donorCondition" value=condition}}
-        {{condition}}
-      </label>
-    {{/each}}
-  </p>
-
-  <p><button {{action "createItem"}}>Next</button></p>
+  <button {{ action 'addDetails'}} disabled class='add_item_link'>Next</button>
 </p>
